Add OptionSet to encapsulate LLVMGEPNoWrapFlags and changed AddressSpace to an Int as it is the correct default Swift

+++ Sources/LLVMSwift/LLVMPointerType.swift
@@ -2,10 +2,10 @@ import LLVM
 
 public class LLVMPointerType<Pointee: LLVMValue>: LLVMType<LLVMPointer<Pointee>> {
     
-    public typealias AddressSpace = UInt32
+    public typealias AddressSpace = Int
     
     @inlinable public convenience init(pointeeType: LLVMType<Pointee>, addressSpace: AddressSpace = 0) {
-        self.init(type: LLVM.LLVMPointerType(pointeeType.type, addressSpace))
+        self.init(type: LLVM.LLVMPointerType(pointeeType.type, UInt32(addressSpace)))
     }
 }
